Fix review issues for Slice 10b bandwidth quotas
Critical fixes:
- Coturn restart handling: implement virtual-negative-baseline approach
  in poller.go to prevent quota bypass on Coturn restart
- Account dashboard UI: fix field names to use account.quota.* structure
- Gate quota poller on cfg.HasTurn() - don't start without TURN config

Important fixes:
- Change PercentageUsed from float64 to int to avoid rendering issues
- Fix DEFAULT_QUOTA_GB default from 100 to 50 in docker-compose.yml
- Fix QUOTA_CHECK_INTERVAL default from 60m to 5m in docker-compose.yml
- Add backfill for existing users in migration (upgrades from 10a)
- Implement pagination in poll() for deployments with >500 users
- Update tests to match int PercentageUsed type

+++ signaling-server/internal/handler/account.go
@@ -15,7 +15,7 @@ type QuotaInfo struct {
 	RemainingGB    float64   `json:"remaining_gb"`
 	PeriodStart    time.Time `json:"period_start"`
 	PeriodEnd      time.Time `json:"period_end"`
-	PercentageUsed float64   `json:"percentage_used"`
+	PercentageUsed int       `json:"percentage_used"`
 }
 
 // AccountResponse represents account information for the dashboard
@@ -55,11 +55,11 @@ func buildQuotaResponse(limitGB, usedGB float64, periodStart, periodEnd time.Tim
 		remainingGB = 0
 	}
 
-	percentageUsed := 0.0
+	percentageUsed := 0
 	if limitGB > 0 {
-		percentageUsed = (usedGB / limitGB) * 100.0
-		if percentageUsed > 100.0 {
-			percentageUsed = 100.0
+		percentageUsed = int((usedGB / limitGB) * 100.0)
+		if percentageUsed > 100 {
+			percentageUsed = 100
 		}
 	}
